pyproject.toml
[project]
name = "alessandra-antigravity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bandit>=1.9.2",
    "cryptography>=46.0.3",
    "django>=5.2.10",
    "django-ninja>=1.5.3",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pytest-django>=4.11.1",
    "uvicorn[standard]>=0.40.0",
    "whitenoise>=6.9.0",
]
